Refactor workshop for use-case and for 2018
Right! So many changes here. This commit essentially refactors most
of the workshop and removes stuff that actually isn't necessary for
the workshop itself, and is only included for the sake of developing
and contributing code. This also removes test validation, since they're
arguably not needed and didn't seem to work atm, as well as the entire
workshop scripts thing which was a bit weird.

This also includes a consistency pass, styling update, and actual update
for changes in React 16.3 related to new and deprecated lifecycle
methods and context API. In truth, this removed all references to
context since that's kind of an advanced feature and because a 2h
workshop will probably end well before reaching that.